+++ Reinvest/LegalEntities/src/Adapter/Database/DatabaseAdapter.ts
@@ -1,5 +1,6 @@
 import { Kysely } from 'kysely';
 import {
+  LegalEntitiesBeneficiary,
   LegalEntitiesCompanyAccount,
   LegalEntitiesDraftAccount,
   LegalEntitiesIndividualAccount,
@@ -11,8 +12,10 @@ export const legalEntitiesProfileTable = 'legal_entities_profile';
 export const legalEntitiesDraftAccountTable = 'legal_entities_draft_accounts';
 export const legalEntitiesIndividualAccountTable = 'legal_entities_individual_account';
 export const legalEntitiesCompanyAccountTable = 'legal_entities_company_account';
+export const legalEntitiesBeneficiaryTable = 'legal_entities_beneficiary';
 
 export interface LegalEntitiesDatabase {
+  [legalEntitiesBeneficiaryTable]: LegalEntitiesBeneficiary;
   [legalEntitiesCompanyAccountTable]: LegalEntitiesCompanyAccount;
   [legalEntitiesDraftAccountTable]: LegalEntitiesDraftAccount;
   [legalEntitiesIndividualAccountTable]: LegalEntitiesIndividualAccount;