+++ Reinvest/LegalEntities/src/Port/Api/ReadAccountController.ts
@@ -5,9 +5,11 @@ import {
   IndividualAccountForSynchronization,
   StakeholderForSynchronization,
 } from 'LegalEntities/Adapter/Database/Repository/AccountRepository';
+import { BeneficiaryRepository } from 'LegalEntities/Adapter/Database/Repository/BeneficiaryRepository';
 import { AccountType } from 'LegalEntities/Domain/AccountType';
 import { AddressInput } from 'LegalEntities/Domain/ValueObject/Address';
 import { DocumentSchema } from 'LegalEntities/Domain/ValueObject/Document';
+import { SensitiveNumberSchema } from 'LegalEntities/Domain/ValueObject/SensitiveNumber';
 import { StakeholderOutput, StakeholderSchema } from 'LegalEntities/Domain/ValueObject/Stakeholder';
 import { AvatarOutput, AvatarQuery } from 'LegalEntities/Port/Api/AvatarQuery';
 
@@ -72,7 +74,7 @@ export type BeneficiaryAccountResponse = {
     };
   };
   id: string;
-  label: string;
+  label?: string;
 };
 
 export type AccountsOverviewResponse = {
@@ -85,10 +87,12 @@ export type AccountsOverviewResponse = {
 export class ReadAccountController {
   private accountRepository: AccountRepository;
   private avatarQuery: AvatarQuery;
+  private beneficiaryRepository: BeneficiaryRepository;
 
-  constructor(accountRepository: AccountRepository, avatarQuery: AvatarQuery) {
+  constructor(accountRepository: AccountRepository, avatarQuery: AvatarQuery, beneficiaryRepository: BeneficiaryRepository) {
     this.accountRepository = accountRepository;
     this.avatarQuery = avatarQuery;
+    this.beneficiaryRepository = beneficiaryRepository;
   }
 
   public static getClassName = (): string => 'ReadAccountController';
@@ -166,6 +170,18 @@ export class ReadAccountController {
       });
     }
 
+    const beneficiaryAccounts = await this.beneficiaryRepository.findBeneficiaryAccounts(profileId);
+
+    for (const beneficiaryAccount of beneficiaryAccounts) {
+      const schema = beneficiaryAccount.toObject();
+      accountsOverview.push(<AccountsOverviewResponse>{
+        id: schema.accountId,
+        type: AccountType.BENEFICIARY,
+        label: schema.label,
+        avatar: await this.avatarQuery.getAvatarForAccount(beneficiaryAccount),
+      });
+    }
+
     return accountsOverview;
   }
 
@@ -197,7 +213,7 @@ export class ReadAccountController {
         stakeholders: accountObject.stakeholders.map(
           (stakeholder: StakeholderSchema): StakeholderOutput => ({
             ...stakeholder,
-            ssn: stakeholder.ssn.anonymized,
+            ssn: (<SensitiveNumberSchema>stakeholder.ssn).anonymized,
             idScan: stakeholder.idScan.map((idScan: DocumentSchema) => ({
               id: idScan.id,
               fileName: idScan.fileName,
@@ -208,26 +224,28 @@ export class ReadAccountController {
     };
   }
 
-  public async readBeneficiaryAccount(profileId: string, accountId: string): Promise<BeneficiaryAccountResponse> {
-    const account = await this.accountRepository.findBeneficiaryAccount(profileId, accountId);
+  public async readBeneficiaryAccount(profileId: string, accountId: string): Promise<BeneficiaryAccountResponse | null> {
+    try {
+      const beneficiaryAccount = await this.beneficiaryRepository.findBeneficiary(profileId, accountId);
 
-    return account;
-    // if (account === null) {
-    //   return {} as BeneficiaryAccountResponse;
-    // }
-    //
-    // const accountObject = account.toObject();
-    //
-    // return {
-    //   id: accountObject.accountId,
-    //   label: account.getLabel(),
-    //   avatar: await this.avatarQuery.getAvatarForAccount(account),
-    //   details: {
-    //     name: {
-    //       firstName: accountObject.firstName,
-    //       lastName: accountObject.lastName,
-    //     },
-    //   },
-    // };
+      if (!beneficiaryAccount) {
+        return null;
+      }
+
+      const schema = beneficiaryAccount.toObject();
+
+      return {
+        avatar: await this.avatarQuery.getAvatarForAccount(beneficiaryAccount),
+        details: {
+          name: schema.name,
+        },
+        id: schema.accountId,
+        label: schema.label,
+      };
+    } catch (error: any) {
+      console.error(error);
+
+      return null;
+    }
   }
 }